c18n: Clear unused registers during compartment transition
Implements unused argument/return value register clearing for dynamic
functions resolved by name (i.e., not function pointers). This
implementation relies on a custom ELF section recording the signature of
each dynamic function.
