+++ libexec/rtld-elf-c18n/Makefile
@@ -4,9 +4,5 @@ MLINKS=	rtld.1 ld-elf-c18n.so.1.1
 
 RTLD_SANDBOX=
 
-.ifdef TRAMP_LINEAR_INSERTION
-CFLAGS+=-DTRAMP_LINEAR_INSERTION
-.endif
-
 .PATH:  ${SRCTOP}/libexec/rtld-elf
 .include "${SRCTOP}/libexec/rtld-elf/Makefile"
